user apdating changes

+++ src/app/components/users/user-profile/user-profile.component.ts
@@ -4,7 +4,7 @@ import { Router } from '@angular/router';
 import { User } from 'src/app/models/User';
 import { AuthService } from 'src/app/services/auth.service';
 import { UsersService } from 'src/app/services/users.service';
-import { ChangesMessageComponent } from '../changes-message/changes-message.component';
+import { WarningDialogComponent } from '../../post/post-dialogs/warning-dialog/warning-gialog.component';
 
 @Component({
   selector: 'app-user-profile',
@@ -13,6 +13,13 @@ import { ChangesMessageComponent } from '../changes-message/changes-message.comp
 })
 export class UserProfileComponent implements OnInit {
   user: User;
+  matDialogConfig = {
+    width: 'auto',
+    height: 'auto',
+    maxHeight: '100vh',
+    maxWidth: '94vw',
+  };
+
   constructor(
     private usersService: UsersService,
     private authService: AuthService,
@@ -24,15 +31,24 @@ export class UserProfileComponent implements OnInit {
     this.user = this.usersService.getCurrentUser();
   }
 
-  updateUser(id: string | undefined, key: string, value: string) {
-    if (id !== undefined) {
-      this.usersService.updateUser(id, key, value);
-      this.usersService.setCurrentUser(this.user);
-      this.dialog.open(ChangesMessageComponent, {
-        width: '30%',
-        height: '30%',
+  updateUserNow(id: string | undefined, user: User) {
+    this.dialog
+      .open(WarningDialogComponent, {
+        data: {
+          title: 'You want to change your name',
+          message: 'Do you confirm the changes?',
+          firstButtonText: 'Cancel',
+          secondButtonText: 'Confirm',
+        },
+        ...this.matDialogConfig,
+      })
+      .afterClosed()
+      .subscribe((result) => {
+        if (result && id !== undefined) {
+          this.usersService.updateUserNow(id, user);
+          this.usersService.setCurrentUser(this.user);
+        }
       });
-    }
   }
 
   updateUserEmail(
@@ -41,27 +57,68 @@ export class UserProfileComponent implements OnInit {
     value: string,
     email: string
   ) {
-    if (id !== undefined) {
-      this.authService.updateEmail(email);
-      this.usersService.updateUser(id, key, value).then(()=> {
-      this.usersService.setCurrentUser(this.user);
-      this.dialog.open(ChangesMessageComponent, {
-        width: '30%',
-        height: '30%',
+    this.dialog
+      .open(WarningDialogComponent, {
+        data: {
+          title: 'You want to change your email',
+          message: 'Do you confirm the changes?',
+          firstButtonText: 'Cancel',
+          secondButtonText: 'Confirm',
+        },
+        ...this.matDialogConfig,
+      })
+      .afterClosed()
+      .subscribe((result) => {
+        if (result) {
+          if (id !== undefined) {
+            this.authService.updateEmail(email);
+            this.usersService.updateUser(id, key, value).then(() => {
+              this.usersService.setCurrentUser(this.user);
+            });
+          }
+        }
       });
-    })
   }
-}
 
   updatePassword(password: string) {
-    this.authService.updatePassword(password);
-    
+    this.dialog
+      .open(WarningDialogComponent, {
+        data: {
+          title: 'You want to change your password',
+          message: 'Do you confirm the changes?',
+          firstButtonText: 'Cancel',
+          secondButtonText: 'Confirm',
+        },
+        ...this.matDialogConfig,
+      })
+      .afterClosed()
+      .subscribe((result) => {
+        if (result) {
+          this.authService.updatePassword(password);
+        }
+      });
   }
 
   deleteUserAccount(key: string | undefined) {
     if (key !== undefined) {
-      this.authService.deleteAccount(key);
-      this.router.navigate(['/'])
+      this.dialog
+
+        .open(WarningDialogComponent, {
+          data: {
+            title: 'You are about to delete your account',
+            message: 'Delete this profile?',
+            firstButtonText: 'Cancel',
+            secondButtonText: 'Confirm',
+          },
+          ...this.matDialogConfig,
+        })
+        .afterClosed()
+        .subscribe((result) => {
+          if (result) {
+            this.authService.deleteAccount(key);
+            this.router.navigate(['/']);
+          }
+        });
     }
   }
 }
